Modified code so that geometry_example actually works.
